fix: download function

+++ evox/lib/net.py
@@ -4,7 +4,7 @@ import lib.log as log
 
 def download(link, file_name, dl_log=True):
     if os.path.exists(link):
-        if not os.path.samefile(link, file_name):
+        if os.path.exists(file_name) and not os.path.samefile(link, file_name):
             shutil.copyfile(link, file_name)
         return
